+++ src/pm/process/manager.rs
@@ -212,7 +212,7 @@ impl ProcessManagerInner {
         )?;
 
         let pid: ProcessIdentifier = self.next_pid;
-        self.next_pid = ProcessIdentifier::from(Into::<usize>::into(pid) + 1);
+        self.next_pid = ProcessIdentifier::from(Into::<u32>::into(pid) + 1);
         let identity: ProcessIdentity = self.get_running().state().identity().clone();
         let process: RunnableProcess = RunnableProcess::new(pid, identity, thread, vmem);
